refactor: 💡 extract outline component
